requirements.txt
faster-whisper
sounddevice
numpy
piper-tts
keyboard
psutil
pyautogui
noisereduce
scipy
pathvalidate
soundfile
pvporcupine
pyaudio
python-docx
PyQt6